requirements.txt
gradio==5.29.1
numpy==2.2.5
openai==1.79.0
openai_whisper==20240930
pyaudio==0.2.14
pydub==0.25.1
Requests==2.32.3
torch==2.7.0+cu118
webrtcvad==2.0.10
